Get stock candles

+++ src/ThreeFourteen.FinnhubClient/Model/RecommendationTrend.cs
@@ -5,22 +5,22 @@ namespace ThreeFourteen.FinnhubClient.Model
     public class RecommendationTrend
     {
         [JsonProperty("buy")]
-        public long Buy { get; set; }
+        public int Buy { get; set; }
 
         [JsonProperty("hold")]
-        public long Hold { get; set; }
+        public int Hold { get; set; }
 
         [JsonProperty("period")]
         public string Period { get; set; }
 
         [JsonProperty("sell")]
-        public long Sell { get; set; }
+        public int Sell { get; set; }
 
         [JsonProperty("strongBuy")]
-        public long StrongBuy { get; set; }
+        public int StrongBuy { get; set; }
 
         [JsonProperty("strongSell")]
-        public long StrongSell { get; set; }
+        public int StrongSell { get; set; }
 
         [JsonProperty("symbol")]
         public string Symbol { get; set; }
